+++ library/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the library will be documented in this file.
 
 - Add support for `special` schema as key of `record` schema (issue #291)
 - Add support for `special` and `specialAsync` schema as key of `recordAsync` schema (issue #291)
+- Add `creditCard` validation function (pull request #292)
 
 ## v0.23.0 (December 08, 2023)